requirements.txt
absl-py==0.12.0
aiogram==2.13
aiohttp==3.7.4
aioredis==1.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
azure-functions==1.7.2
Babel==2.9.1
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
emoji==1.2.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.2
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
hiredis==2.0.0
idna==2.10
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.3.4
multidict==5.1.0
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tokenizers==0.10.3
tqdm==4.61.0
typing-extensions==3.7.4.3
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
